browser router and component refactor
