+++ src/components/App/App.js
@@ -1,19 +1,19 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { exchangeToken, logout } from './state/actionCreators/authAC';
-import { fetchCart } from './state/actionCreators/cartAC';
-import { Route, Routes, NavLink, HashRouter } from 'react-router-dom';
-import Cart from './components/Cart';
-import Home from './components/Home';
-import Login from './components/Login';
-import Products from './components/Products';
-import User from './components/User';
-import Product from './components/Product';
+import React, {useEffect} from 'react';
+import {useDispatch, useSelector} from 'react-redux';
+import {exchangeToken, logout} from '../../state/actionCreators/authAC';
+import {fetchCart} from '../../state/actionCreators/cartAC';
+import {Route, Routes, NavLink} from 'react-router-dom';
+import Cart from '../Cart';
+import Home from '../Home';
+import Login from '../Login';
+import Products from '../Products';
+import User from '../User';
+import Product from '../Product';
 
 function App() {
     const dispatch = useDispatch();
-    const { auth } = useSelector((state) => state.auth);
-    const { cart } = useSelector((state) => state.cart);
+    const {auth} = useSelector((state) => state.auth);
+    const {cart} = useSelector((state) => state.cart);
     const fetchData = () => {
         dispatch(exchangeToken());
         if (auth.id) dispatch(fetchCart());
@@ -27,7 +27,7 @@ function App() {
         if (auth.id) dispatch(fetchCart());
     }, [auth]);
     return (
-        <HashRouter>
+        <div id='app'>
             <div className="nav-bar">
                 <NavLink to="/">LOGO</NavLink>
                 <NavLink to="/products/category/A">CategoryA</NavLink>
@@ -45,26 +45,27 @@ function App() {
                 )}
             </div>
             <Routes>
-                <Route exact path="/" element={<Home />} />
+                <Route exact path="/" element={<Home/>}/>
                 <Route
                     path="/products/category/A"
-                    element={<Products categoryId="1" />}
+                    element={<Products categoryId="1"/>}
                 />
                 <Route
                     path="/products/category/B"
-                    element={<Products categoryId="2" />}
+                    element={<Products categoryId="2"/>}
                 />
                 <Route
                     path="/products/category/C"
-                    element={<Products categoryId="3" />}
+                    element={<Products categoryId="3"/>}
                 />
-                <Route path="/user" element={<User />} />
-                <Route path="/login" element={<Login />} />
-                <Route path="/cart" element={<Cart />} />
-                <Route path="/products/:productId" element={<Product />} />
+                <Route path="/user" element={<User/>}/>
+                <Route path="/login" element={<Login/>}/>
+                <Route path="/cart" element={<Cart/>}/>
+                <Route path="/products/:productId" element={<Product/>}/>
             </Routes>
-        </HashRouter>
-    );
+        </div>
+    )
+        ;
 }
 
 export default App;
